Get AWS auth token for each ECR registry
Instead of asking for the AWS region and set of registry IDs (really
account IDs), take these from the registry URLs as they come up.

This means we have to make a token request for each region; so, keep
track of the expiry, per region. And so we don't keep asking when
there's a failure, embargo requests for a region (for ten minutes)
whenever we fail to get a token.